requirements/dev.txt
uvicorn~=0.32.1
fastapi-decorators~=1.0.6
python-dotenv~=1.0.1
alembic~=1.14.1
python-multipart~=0.0.26
ultralytics~=8.4.37
opencv-python-headless~=4.13.0
